마이페이지는 로그인을 하셔야한다는 alert 빼고, MyPageSection2에서는 이메일 변경 컴포넌트로 빼줬고, 이메일 변경로직은 이메일 형식으로 입력해야만 바뀌도록 함
